Rename `FeedViewController` To `FeedUIIntegrationTests` since it tests multiple UI components
