+++ EssentialFeediOSTests/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,18 @@
+//
+//  UIRefreshControl+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Macbook Pro on 20/9/23.
+//
+
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        allTargets.forEach({ target in
+            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach {
+                (target as NSObject).perform(Selector($0))
+            }
+        })
+    }
+}
